Moved GridSpace/NdLayout overlay implementation to baseclass
